backend/requirements.txt
fastapi
uvicorn
openai
fpdf
smtplib
email-validator
python-dotenv
